requirements.txt
fastapi
uvicorn
pymongo
requests
pydantic
fuzzywuzzy
python-dotenv
